+++ src/main/java/insane96mcp/iguanatweaksreborn/module/world/Sextant.java
@@ -4,7 +4,6 @@ import insane96mcp.iguanatweaksreborn.module.Modules;
 import insane96mcp.iguanatweaksreborn.setup.ISORegistries;
 import insane96mcp.insanelib.base.Feature;
 import insane96mcp.insanelib.base.LoadFeature;
-import insane96mcp.insanelib.base.Module;
 import net.minecraft.core.Holder;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.entity.Entity;
@@ -16,12 +15,8 @@ import net.minecraft.world.level.biome.Biome;
 import net.minecraftforge.registries.RegistryObject;
 
 @LoadFeature(module = Modules.Ids.WORLD, canBeDisabled = false)
-public class BiomeCompass extends Feature {
-    public static final RegistryObject<Item> COMPASS = ISORegistries.ITEMS.register("biome_compass", () -> new BiomeCompassItem(new Item.Properties()));
-
-    public BiomeCompass(Module module, boolean enabledByDefault, boolean canBeDisabled) {
-        super(module, enabledByDefault, canBeDisabled);
-    }
+public class Sextant extends Feature {
+    public static final RegistryObject<Item> ITEM = ISORegistries.ITEMS.register("sextant", () -> new BiomeCompassItem(new Item.Properties()));
 
     public static class BiomeCompassItem extends Item {
         public BiomeCompassItem(Properties pProperties) {